added experiment iteration

+++ scripts/linux/merge-udp-test.sh
@@ -2,6 +2,7 @@
 PPS=$1
 BYTE=$2
 N=$3
+RUN_TEST_SAMENODE=$4
 BASE_FOLDER=/vagrant/ext/kites/pod-shared
 cd $BASE_FOLDER
 echo -e "NETSNIFF TEST - ${BYTE}byte - ${PPS}pps\n" > NETSNIFF-${BYTE}byte-${PPS}pps.txt
@@ -24,15 +25,17 @@ do
     rm temp_trafgen.txt
 done
 
-cd $BASE_FOLDER/single-pod
-cat NETSNIFF-${BYTE}byte-${PPS}pps.txt > temp_netsniff.txt
-mv temp_netsniff.txt ..
-#rm NETSNIFF-${BYTE}byte-${PPS}pps.txt
-cat TRAFGEN-${BYTE}byte-${PPS}pps.txt > temp_trafgen.txt
-mv temp_trafgen.txt ..
-#rm TRAFGEN-${BYTE}byte-${PPS}pps.txt
-cd .. 
-cat temp_netsniff.txt >> NETSNIFF-${BYTE}byte-${PPS}pps.txt
-rm temp_netsniff.txt
-cat temp_trafgen.txt >> TRAFGEN-${BYTE}byte-${PPS}pps.txt
-rm temp_trafgen.txt
+if $RUN_TEST_SAMENODE; then
+    cd $BASE_FOLDER/single-pod
+    cat NETSNIFF-${BYTE}byte-${PPS}pps.txt > temp_netsniff.txt
+    mv temp_netsniff.txt ..
+    #rm NETSNIFF-${BYTE}byte-${PPS}pps.txt
+    cat TRAFGEN-${BYTE}byte-${PPS}pps.txt > temp_trafgen.txt
+    mv temp_trafgen.txt ..
+    #rm TRAFGEN-${BYTE}byte-${PPS}pps.txt
+    cd .. 
+    cat temp_netsniff.txt >> NETSNIFF-${BYTE}byte-${PPS}pps.txt
+    rm temp_netsniff.txt
+    cat temp_trafgen.txt >> TRAFGEN-${BYTE}byte-${PPS}pps.txt
+    rm temp_trafgen.txt
+fi
